colors work, need to fix zoom

+++ full_libft/include/libft.h
@@ -94,7 +94,7 @@ int					ft_lstsize(t_list *lst);
 */
 int					count_words(char const *s, char c);
 int					ft_abs(int nb);
-int					ft_atol(const char *nptr);
+long				ft_atol(const char *nptr);
 int					ft_strcmp(const char *s1, const char *s2);
 void				ft_free_split(char	**split);
 int					ft_atoi_base(char *nbr, char *base_from);
